Refactor!: Moved scheduler to its own module pproc.c should just contain the CLI
